+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [STVP](https://mia.kiwi/projects/stvp).
 
+## [25.0.7] - 2025-08-26
+
+### Added
+
+- App configuration file
+- Special metadata attribute 'related_posts' to suggest posts related to the current one
+- Support for HTML `meta` tags
+- Dynamic page title generation
+- Special text when viewing an unlisted post
+
+### Changed
+
+- Footer links and notice can be configured in the app settings file
+- The `getPostCardBody`, `getPostCard`, and `showPostCard` methods of Post are now async
+- Posts are now sorted by update time first, then publication time. Newly updated posts will now be pushed back to the start of the posts container
+
+
+
 ## [25.0.6] - 2025-08-21
 
 ### Added